requirements.txt
Flask==3.0.0
joblib
python-dotenv